requirements.txt
gemini-python==0.2.0